Move app tray outside of main.js.
